+++ setup.py
@@ -35,6 +35,7 @@ setup(
         'django-next-url-mixin>=0.1.0',
         'suds',
         'django-sequences',
+        'swapper',
         'six',
     ],
     extras_require={
